♻️ api: rename src certification models directory
