+++ api/src/certification/flash-certification/domain/models/AssessmentSimulatorSingleMeasureStrategy.js
