+++ api/src/certification/flash-certification/domain/models/FlashAssessmentAlgorithmConfiguration.js
