+++ api/src/certification/flash-certification/domain/models/FlashAssessmentSuccessRateHandlerFixedStrategy.js
