+++ api/tests/certification/flash-certification/unit/domain/models/FlashAssessmentAlgorithmForcedCompetencesRule_test.js
@@ -1,4 +1,4 @@
-import { FlashAssessmentAlgorithmForcedCompetencesRule } from '../../../../../../src/certification/flash-certification/domain/model/FlashAssessmentAlgorithmForcedCompetencesRule.js';
+import { FlashAssessmentAlgorithmForcedCompetencesRule } from '../../../../../../src/certification/flash-certification/domain/models/FlashAssessmentAlgorithmForcedCompetencesRule.js';
 import { domainBuilder, expect } from '../../../../../test-helper.js';
 
 describe('Unit | Domain | Models | FlashAssessmentAlgorithm | FlashAssessmentAlgorithmForcedCompetencesRule', function () {
