+++ api/tests/certification/flash-certification/unit/domain/models/FlashAssessmentAlgorithmOneQuestionPerTubeRule_test.js
@@ -1,4 +1,4 @@
-import { FlashAssessmentAlgorithmOneQuestionPerTubeRule } from '../../../../../../src/certification/flash-certification/domain/model/FlashAssessmentAlgorithmOneQuestionPerTubeRule.js';
+import { FlashAssessmentAlgorithmOneQuestionPerTubeRule } from '../../../../../../src/certification/flash-certification/domain/models/FlashAssessmentAlgorithmOneQuestionPerTubeRule.js';
 import { domainBuilder, expect } from '../../../../../test-helper.js';
 
 describe('Unit | Domain | Models | FlashAssessmentAlgorithm | FlashAssessmentAlgorithmOneQuestionPerTubeRule', function () {
