+++ api/tests/certification/flash-certification/unit/domain/models/FlashAssessmentAlgorithmRuleEngine_test.js
@@ -1,5 +1,5 @@
 import { expect } from '../../../../../test-helper.js';
-import { FlashAssessmentAlgorithmRuleEngine } from '../../../../../../src/certification/flash-certification/domain/model/FlashAssessmentAlgorithmRuleEngine.js';
+import { FlashAssessmentAlgorithmRuleEngine } from '../../../../../../src/certification/flash-certification/domain/models/FlashAssessmentAlgorithmRuleEngine.js';
 
 const challenge1 = Symbol('challenge1');
 const challenge2 = Symbol('challenge2');
