+++ api/tests/certification/flash-certification/unit/domain/models/FlashAssessmentSuccessRateHandler_test.js
@@ -1,4 +1,4 @@
-import { FlashAssessmentSuccessRateHandler } from '../../../../../../src/certification/flash-certification/domain/model/FlashAssessmentSuccessRateHandler.js';
+import { FlashAssessmentSuccessRateHandler } from '../../../../../../src/certification/flash-certification/domain/models/FlashAssessmentSuccessRateHandler.js';
 import { expect } from '../../../../../test-helper.js';
 
 describe('Unit | Domain | Models | FlashAssessmentAlgorithmSuccessRateHandler', function () {
